requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
groq==0.4.1
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0